requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai-whisper==20231117
torch
ollama==0.2.1
tensorflow==2.16.1
tensorflow-hub==0.16.1
streamlit==1.35.0
python-multipart==0.0.9
sqlalchemy==2.0.30
aiofiles==23.2.1
python-dotenv==1.0.1
langdetect==1.0.9
reportlab==4.1.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
httpx==0.27.0
Pillow==10.3.0
